Adding the null check for medicaitonmap id
git-tfs-id: [http://hillstfs2013:8080/tfs/Phytel]$/PhytelCode/Phytel.Net/Services;C13830

+++ API/DataDomain/Phytel.API.DataDomain.Medication/Development/Current/Medication/MedicationMappingDataManager.cs
@@ -76,8 +76,11 @@ namespace Phytel.API.DataDomain.Medication
                     request.MedicationMaps.ForEach(m =>
                     {
                         string id = (string)repo.Find(m);
-                        m.Id = id;
-                        repo.Delete(id);
+                        if (!string.IsNullOrEmpty(id))
+                        {
+                            m.Id = id;
+                            repo.Delete(id);
+                        }
                     });
                 }
                 return request.MedicationMaps;
